+++ hqc/gui.py
