+++ hqc/hqc.py
