+++ hqc/sample_conn.conf
